requirements.txt
adafruit_circuitpython_dht==3.7.4
adafruit_dht==1.4.0
paho_mqtt==1.6.1
